+++ back/src/main/java/com/cocktailpick/back/common/csv/CsvReader.java
@@ -1,0 +1,7 @@
+package com.cocktailpick.back.common.csv;
+
+import java.util.List;
+
+public interface CsvReader {
+	List<String[]> readAll();
+}
