+++ back/src/main/java/com/cocktailpick/back/tag/service/TagCsvReader.java
@@ -1,15 +1,13 @@
 package com.cocktailpick.back.tag.service;
 
-import java.io.IOException;
-import java.io.InputStreamReader;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import org.springframework.web.multipart.MultipartFile;
 
+import com.cocktailpick.back.common.csv.CsvReader;
+import com.cocktailpick.back.common.csv.OpenCsvReader;
 import com.cocktailpick.back.tag.domain.Tag;
-import com.opencsv.CSVReader;
-import com.opencsv.CSVReaderBuilder;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 
@@ -17,34 +15,16 @@ import lombok.RequiredArgsConstructor;
 public class TagCsvReader {
 	private static final int NAME_COLUMN_INDEX = 0;
 
-	private final CSVReader csvReader;
+	private final CsvReader csvReader;
 
 	public static TagCsvReader from(MultipartFile file) {
-		return new TagCsvReader(createCsvReader(file));
-	}
-
-	private static CSVReader createCsvReader(MultipartFile file) {
-		return new CSVReaderBuilder(createInputStreamReader(file))
-			.withSkipLines(1)
-			.build();
-	}
-
-	private static InputStreamReader createInputStreamReader(MultipartFile file) {
-		try {
-			return new InputStreamReader(file.getInputStream());
-		} catch (IOException e) {
-			throw new RuntimeException();
-		}
+		return new TagCsvReader(OpenCsvReader.from(file));
 	}
 
 	public List<Tag> getTags() {
-		try {
-			List<String[]> tags = csvReader.readAll();
-			return tags.stream()
-				.map(line -> new Tag(line[NAME_COLUMN_INDEX]))
-				.collect(Collectors.toList());
-		} catch (IOException e) {
-			throw new RuntimeException();
-		}
+		List<String[]> tags = csvReader.readAll();
+		return tags.stream()
+			.map(line -> new Tag(line[NAME_COLUMN_INDEX]))
+			.collect(Collectors.toList());
 	}
 }
